Flytter Loader til root.tsx, Termliste til termliste.tsx
