+++ app/lib/components/loader.tsx
@@ -1,7 +1,7 @@
 import '~/styles/loader.css';
 
 export const Loader = () => (
-  <div className="loader book">
+  <div className="loader book" role="status" aria-label="Laster innhold">
     <div className="page"></div>
     <div className="page"></div>
     <div className="page"></div>
